+++ hoot-core/src/main/cpp/hoot/core/algorithms/WayJoiner2.cpp
@@ -28,7 +28,7 @@
 #include "WayJoiner2.h"
 
 //  Hoot
-#include <hoot/core/conflate/NodeToWayMap.h>
+#include <hoot/core/elements/NodeToWayMap.h>
 #include <hoot/core/index/OsmMapIndex.h>
 #include <hoot/core/ops/RecursiveElementRemover.h>
 #include <hoot/core/ops/ReplaceElementOp.h>
