+++ hoot-core/src/main/cpp/hoot/core/algorithms/splitter/LargeWaySplitter.cpp
@@ -34,7 +34,7 @@
 #include <hoot/core/elements/OsmMap.h>
 #include <hoot/core/algorithms/splitter/WaySplitter.h>
 #include <hoot/core/algorithms/linearreference/WayLocation.h>
-#include <hoot/core/conflate/NodeToWayMap.h>
+#include <hoot/core/elements/NodeToWayMap.h>
 #include <hoot/core/elements/Way.h>
 #include <hoot/core/index/OsmMapIndex.h>
 #include <hoot/core/util/ConfigOptions.h>
