+++ hoot-core/src/main/cpp/hoot/core/conflate/highway/HighwayMergerAbstract.cpp
@@ -36,7 +36,7 @@
 #include <hoot/core/util/MapProjector.h>
 #include <hoot/core/algorithms/subline-matching/SublineStringMatcher.h>
 #include <hoot/core/algorithms/splitter/MultiLineStringSplitter.h>
-#include <hoot/core/conflate/NodeToWayMap.h>
+#include <hoot/core/elements/NodeToWayMap.h>
 #include <hoot/core/conflate/highway/HighwayMatch.h>
 #include <hoot/core/index/OsmMapIndex.h>
 #include <hoot/core/io/OsmMapWriterFactory.h>
