+++ hoot-core/src/main/cpp/hoot/core/conflate/highway/Roundabout.cpp
@@ -28,7 +28,7 @@
 #include "Roundabout.h"
 #include <hoot/core/util/MapProjector.h>
 #include <hoot/core/io/OsmMapWriterFactory.h>
-#include <hoot/core/conflate/NodeToWayMap.h>
+#include <hoot/core/elements/NodeToWayMap.h>
 #include <hoot/core/index/OsmMapIndex.h>
 #include <hoot/core/ops/RemoveWayOp.h>
 #include <hoot/core/ops/RemoveNodeOp.h>
