+++ hoot-core/src/main/cpp/hoot/core/conflate/matching/NodeMatcher.cpp
@@ -29,7 +29,7 @@
 
 // Hoot
 #include <hoot/core/algorithms/WayHeading.h>
-#include <hoot/core/conflate/NodeToWayMap.h>
+#include <hoot/core/elements/NodeToWayMap.h>
 #include <hoot/core/index/OsmMapIndex.h>
 #include <hoot/core/util/ConfigOptions.h>
 #include <hoot/core/util/MapProjector.h>
