+++ hoot-core/src/main/cpp/hoot/core/conflate/network/NetworkMerger.cpp
@@ -31,7 +31,7 @@
 #include <hoot/core/algorithms/WayMatchStringMerger.h>
 #include <hoot/core/index/OsmMapIndex.h>
 #include <hoot/core/io/OsmJsonWriter.h>
-#include <hoot/core/conflate/NodeToWayMap.h>
+#include <hoot/core/elements/NodeToWayMap.h>
 #include <hoot/core/conflate/highway/HighwayMatch.h>
 #include <hoot/core/conflate/review/ReviewMarker.h>
 #include <hoot/core/ops/RecursiveElementRemover.h>
