+++ hoot-core/src/main/cpp/hoot/core/criterion/UselessElementCriterion.cpp
@@ -33,7 +33,7 @@
 #include <hoot/core/elements/Tags.h>
 #include <hoot/core/util/Log.h>
 #include <hoot/core/elements/Element.h>
-#include <hoot/core/conflate/NodeToWayMap.h>
+#include <hoot/core/elements/NodeToWayMap.h>
 
 namespace hoot
 {
