+++ hoot-core/src/main/cpp/hoot/core/ops/RemoveNodeOp.cpp
@@ -29,7 +29,7 @@
 // hoot
 #include <hoot/core/util/Factory.h>
 #include <hoot/core/index/OsmMapIndex.h>
-#include <hoot/core/conflate/NodeToWayMap.h>
+#include <hoot/core/elements/NodeToWayMap.h>
 #include <hoot/core/util/Validate.h>
 
 using namespace std;
