+++ hoot-core/src/main/cpp/hoot/core/visitors/FindHighwayIntersectionsVisitor.cpp
@@ -28,7 +28,7 @@
 
 #include <hoot/core/elements/OsmMap.h>
 #include <hoot/core/index/OsmMapIndex.h>
-#include <hoot/core/conflate/NodeToWayMap.h>
+#include <hoot/core/elements/NodeToWayMap.h>
 #include <hoot/core/conflate/matching/NodeMatcher.h>
 #include <hoot/core/util/Factory.h>
 #include <hoot/core/criterion/HighwayCriterion.h>
