+++ hoot-core/src/main/cpp/hoot/core/visitors/ImplicitTypeTaggerBase.cpp
@@ -33,7 +33,7 @@
 #include <hoot/core/util/ConfigOptions.h>
 #include <hoot/core/util/StringUtils.h>
 #include <hoot/core/schema/ImplicitTagUtils.h>
-#include <hoot/core/conflate/poi-polygon/extractors/PoiPolygonTypeScoreExtractor.h>
+#include <hoot/core/algorithms/extractors/poi-polygon/PoiPolygonTypeScoreExtractor.h>
 #include <hoot/core/util/Factory.h>
 
 // Qt
